Separate generic-intro into two distinct includes
Until now there has been a `generic-intro` partial that includes
an optional prefix. This isn't goint to work for those pages that
rely upon this partial but which are not content managed. An exmaple
of this is the record disambiguation page.

Since this is an optional element, rather than a modifier, it seems
most sensible for two templates to exist and not to consider one to
be a modified version of the other.