Add GTM JavaScript to Wagtail repo and refactor
This commit introduces all the GTM tracking code that had previously
been stored in a separate repo. The reason for this change is that
we've been informed by the analytics team that their previous approach
of adding scripts via the GTM console is no longer preferred and they
would like us instead to include the scripts within the main repo.