Add CSS to provide some spacing around XSLT generated markup
We were asked by the data team to provide add CSS in order to
improve the presentation of XLST generated markup which appears
within the Record Description block of details pages. This included
a request to make spans appear to be list items.

I have explained that using CSS to make HTML that lacks appropriate
semantics appear as it would if the correct semantics were in place
would fail WCAG SC 1.3.1. and that the best I can do is to introduce
some padding.

It has been acknowledged by data colleagues that this is an issue
which needs to be addressed at source.