Add <h2> to Explorer Index Page
HeadingsMap revealed non-sequentially descending headings on this
page. My initial thought was to split the ParagraphWithHeading block
into two separate blocks (one for use where a `<h2>` is needed and
one for use where a `<h3>` is needed) but I'm advised that this is
not advisable from a Wagtail perspective. I've therefore added a
`<h2>` into the template.