Add aria-label to audio elements referencing transcript
This is to sure up adherence to SC 1.2.1. by implementing a programmatic
relationship between the `<audio>` element and the transcript which follows.

The G158 Sufficient Technique says:

"Check that... The transcript itself can be programmatically determined
from the text alternative for the audio-only content". Since the `alt`
attribute isn't permitted on `<audio>` elements I have used `aria-label`.