Convert RecordPage to dataclass.
The original intention for a RecordPage was to be able to use it
anywhere a wagtail.core.Page is used. To pass it to a view, for it to
share the same base templates and includes as a standard page instance
and to be selectable in the admin using a custom model picker.

All these goals are achievable without subclassing `wagtail.core.Page`
and since RecordPage data is never fetched from Wagtail, and the page
isn't served using `Page.serve`, there's no need for it to inherit from
wagtail.core.Page.

Our RecordPage instance is a container for Elasticsearch data
served by Kong and so it's more appropriate to use a dataclass.